part 2

+++ libft/srcs/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: dokwak <dokwak@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/30 21:57:01 by dokwak            #+#    #+#             */
-/*   Updated: 2021/11/30 22:55:21 by dokwak           ###   ########.fr       */
+/*   Updated: 2021/12/01 11:53:10 by dokwak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
@@ -14,7 +14,7 @@
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*pnt;
-	int		i;
+	size_t	i;
 
 	i = -1;
 	pnt = malloc(sizeof(char) * (len + 1));
